// zeytinus/ui/base/context/types.d.ts
export interface BaseData {
  csrf_token?: string;
  full_name?: string;
  short_name?: string;
  motto?: string;
  theme_color?: string;
  colored_logo_full_image?: string;
  colored_logo_mini_image?: string;
  light_logo_full_image?: string;
  light_logo_mini_image?: string;
  dark_logo_full_image?: string;
  dark_logo_mini_image?: string;
  primary_phone?: string;
  secondary_phone?: string;
  primary_email?: string;
  secondary_email?: string;
  city_name?: string;
  PO_box?: string;
  street?: string;
  open_days?: string;
  open_hours?: string;
  map?: string;
  social_media_links_version?: "V1" | "V2" | "V3";
  facebook?: string;
  instagram?: string;
  twitter_x?: string;
  linkedin?: string;
  spotify?: string;
}

export interface BaseMetadata {
  full_name?: string;
  motto?: string;
  website?: string;
  favicon?: string;
  apple_touch_icon?: string;
  og_image?: string;
  twitter_image?: string;
}